refactor: NearbyMineCounter 패키지 이동
